refactor(js): extract card functions to `components/card.js` module
Steps:
- Create `assets/js/components/card.js` with `cardCreate`, `cardMoveImage` and `cardWhenClicked` exports
- Keep `cardCreateList`, `cardCreateItem` and `cardToggleView` as internal functions (not exported)
- Import `firstUpperCase` from `utils/first-upper-case.js` in the new module
- Remove card functions from `scripts.js`
- Remove unused `firstUpperCase` import from `scripts.js`
- Add `import { cardCreate, cardMoveImage, cardWhenClicked }` to `scripts.js`

Reason:
- Part of the modularization effort to split `scripts.js` into separate ES modules

References:
- Issue #104
